+++ common/common-custom-security/src/main/java/com/meysam/common/customsecurity/repository/AdminPermissionRepository.java
@@ -7,20 +7,19 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 public interface AdminPermissionRepository extends BaseRepository<AdminPermission> {
 
-    Optional<AdminPermission> findByPermission(BigInteger permission);
+    Optional<AdminPermission> findByPermission(Long permission);
 
-    List<AdminPermission> findAllByAdmin(BigInteger admin);
+    List<AdminPermission> findAllByAdmin(Long admin);
 
     @Query(value = "select p.name from Permission p where p.id in (select ap.permission from AdminPermission ap where ap.admin =:adminId)")
-    List<String> findPermissionsNamesByAdmin(@Param("adminId") BigInteger adminId);
+    List<String> findPermissionsNamesByAdmin(@Param("adminId") Long adminId);
 
-    Optional<AdminPermission> findByAdminAndPermission(BigInteger admin, BigInteger permission);
+    Optional<AdminPermission> findByAdminAndPermission(Long admin, Long permission);
 
 }
