+++ common/common-custom-security/src/main/java/com/meysam/common/customsecurity/service/AdminRoleService.java
@@ -21,7 +21,7 @@ public class AdminRoleService {
         return adminRoleRepository.save(adminRole);
     }
 
-    public List<Role> getRoles(BigInteger profileId) {
+    public List<Role> getRoles(Long profileId) {
         return adminRoleRepository.findRolesByProfileId(profileId);
     }
 }
